Update URL when scrollspy updates navigation
scrollspy updates currently active navigation element based on the
scroll position of user on page. Also update the hash fragment of the
URL when this happens.

Concern:

This is desirable when someone wants to share the exact location on the
page where they are, but may not be desirably when someone's simply
trying to share the link of the site.

We'll remove it if it doesn't work out.