Update member data
Based on two sources: reponses to "taking stock" and people currently
added to mailing list

Removed all 2014-17 links from the old table but left the earlier ones.
We should perhaps rename those as "byld alumni" and have a separate
section?

Then added only the names present in the sources.

One list was missing data for names so I've just used the email as it
was already common in the data which did have names

The other list was missing data for skills and github username, so I
just omitted those fields for now